% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.oc.pbf.R
\name{get.oc.pbf}
\alias{get.oc.pbf}
\title{Generate operating characteristics for single agent trials}
\usage{
get.oc.pbf(target,n.cohort,cohortsize,skeleton,n.trial,
                     risk.cutoff,earlyterm,start)
}
\arguments{
\item{target}{the target DLT rate}

\item{cohortsize}{the cohort size}

\item{skeleton}{a vector containing the true toxicity probabilities of the
investigational dose levels.}

\item{n.trial}{the total number of trials to be simulated}

\item{risk.cutoff}{the cutoff to eliminate an over/under toxic dose.
We recommend the default value of (\code{risk.cutoff=0.8}) for general use.}

\item{earlyterm}{the early termination parameter.}

\item{ncohort}{the total number of cohorts}

\item{seed}{the random seed for simulation}
}
\value{
\code{get.oc()} returns the operating characteristics of the PBF design as a list,
       including:
       (1) selection percentage at each dose level (\code{$selpercent}),
       (2) the number of patients treated at each dose level (\code{$npatients}),
       (3) the number of toxicities observed at each dose level (\code{$ntox}),
       (4) the average number of toxicities (\code{$totaltox}),
       (5) the average number of patients (\code{$totaln}),
       (6) the percentage of early stopping without selecting the MTD (\code{$percentstop}),
       (7) risk of overdosing 60\% or more of patients (\code{$overdose60}),
       (8) risk of overdosing 80\% or more of patients (\code{$overdose80}),
       (9) data.frame (\code{$simu.setup}) containing simulation parameters, such as target, p.true, etc.
}
\description{
Obtain the operating characteristics of the BOIN design for single agent trials by simulating trials.
}
\details{
TBD
}
\note{
TBD
}
\examples{

## get the operating characteristics for BOIN single agent trial
oc <- get.oc.pbf(target=0.3,n.cohort=10,cohortsize=3,skeleton=c(0.3,0.4,0.5,0.6),n.trial=1000,
                     risk.cutoff=0.8,earlyterm=TRUE,start=1)

summary(oc) # summarize design operating characteristics
plot(oc)   # plot flowchart of the BOIN design and design operating characteristics, including
           # selection percentage, number of patients, and observed toxicities at each dose


}
\references{

}
\seealso{

}
\author{

}
