% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.oc.pop.R
\name{get.oc.pop}
\alias{get.oc.pop}
\title{Generate operating characteristics for single agent trials}
\usage{
get.oc.pop(target,n.cohort,cohortsize,titration,skeleton,n.trial,
                     risk.cutoff,earlyterm,start)
}
\arguments{
\item{target}{the target DLT rate}

\item{n.cohort}{the total number of cohorts}

\item{cohortsize}{the cohort size}

\item{titration}{default is TRUE. Set \code{titration=TRUE} to perform dose
escalation with cohort size = 1 to accelerate dose escalation
at the beginning of the trial.}

\item{skeleton}{a vector containing the true toxicity probabilities of the
investigational dose levels.}

\item{n.trial}{the total number of trials to be simulated}

\item{risk.cutoff}{the cutoff to eliminate an over/under toxic dose.
We recommend the default value of (\code{risk.cutoff=0.8}) for general use.}

\item{earlyterm}{the early termination parameter.}

\item{start}{specify the starting dose level. Default value is 1.}
}
\value{
\code{get.oc.pop()} returns the operating characteristics of the PoP design as a list,
       including:
       (1) selection percentage at each dose level (\code{$sel.pct}),
       (2) the number of patients treated at each dose level (\code{$num.p}),
       (3) the number of toxicities observed at each dose level (\code{$num.tox}),
       (4) the average number of toxicities,
       (5) the average number of patients,
       (6) the percentage of early stopping without selecting the MTD (\code{$early}),
       (7) risk of underdosing 80\% or more of patients (\code{$risk.under}),
       (8) risk of overdosing 80\% or more of patients (\code{$risk.over})
}
\description{
Obtain the operating characteristics of the PoP design for single agent trials by simulating trials.
}
\details{
TBD
}
\note{
TBD
}
\examples{

## get the operating characteristics for BOIN single agent trial
oc <- get.oc.pop(target=0.3,n.cohort=10,cohortsize=3,titration=TRUE,
                 skeleton=c(0.3,0.4,0.5,0.6),n.trial=1000,
                     risk.cutoff=0.8,earlyterm=TRUE,start=1)

summary(oc) # summarize design operating characteristics
plot(oc)

}
