% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pbf.R
\name{print.pbf}
\alias{print.pbf}
\title{Generate descriptive summary for objects returned by other functions}
\usage{
\method{print}{pbf}(x, ...)
}
\arguments{
\item{x}{the object returned by other functions}

\item{...}{ignored arguments}
}
\value{
\code{print()} prints the objects returned by other functions.
}
\description{
Generate descriptive summary for objects returned by other functions.
}
\details{
\code{print()} prints the objects returned by other functions.
}
\examples{



}
\author{

}
