% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pop.R
\name{summary.pop}
\alias{summary.pop}
\title{Generate descriptive summary for objects returned by other functions}
\usage{
\method{summary}{pop}(object, ...)
}
\arguments{
\item{object}{the object returned by other functions.}

\item{...}{ignored arguments}
}
\value{
\code{summary()} prints the objects returned by other functions.
}
\description{
Generate descriptive summary for objects returned by other functions.
}
\details{
\code{summary()} prints the objects returned by other functions.
}
